requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
requests>=2.31.0
httpx>=0.27.0
python-dotenv>=1.0.0
mcp>=1.0.0
